adding plugin docstrings
